stocktech-backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pymongo>=3.12.0
motor>=2.6.0
pydantic~=2.4.2
passlib~=1.7.4
PyJWT~=1.7.1
pydantic[email]
gtts
translate